♻️ [refactor] #152 : domain 폴더로 모두 이동
단순 폴더 이동 drag&drop
swagger로 이상 없음 확인
