+++ server/src/domain/alarms/alarms.controller.ts
