+++ server/src/domain/exercises/exercises.module.ts
