+++ server/src/domain/exercises/exercises.service.ts
