+++ server/src/domain/oauth/google-oauth/google-oauth.controller.ts
@@ -3,7 +3,7 @@ import { AuthGuard } from "@nestjs/passport";
 import { Response } from "express";
 import { ApiOperation, ApiTags } from "@nestjs/swagger";
 import { GoogleOauthService } from "./google-oauth.service";
-import { RequestWithUser } from "../../types/request";
+import { RequestWithUser } from "../../../types/request";
 
 @Controller("api/oauth/google")
 @ApiTags("OAUTH API")
