+++ server/src/domain/oauth/google-oauth/google-oauth.module.ts
@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { User } from "src/users/entities/user.entity";
+import { User } from "src/domain/users/entities/user.entity";
 import { JwtModule } from "@nestjs/jwt";
 import { GoogleOauthController } from "./google-oauth.controller";
 import { GoogleStrategy } from "./utils/google.strategy";
 import { GoogleOauthService } from "./google-oauth.service";
 import { JwtStrategy } from "../jwt/jwt.strategy";
-import { ACCESS_TOKEN_EXPIRESIN, ACCESS_TOKEN_SECRETKEY } from "../../utils/env";
+import { ACCESS_TOKEN_EXPIRESIN, ACCESS_TOKEN_SECRETKEY } from "../../../utils/env";
 
 @Module({
   imports: [
