+++ server/src/domain/oauth/google-oauth/google-oauth.service.ts
@@ -1,9 +1,9 @@
 import { BadRequestException, Injectable, InternalServerErrorException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { User } from "src/users/entities/user.entity";
+import { User } from "src/domain/users/entities/user.entity";
 import { JwtService } from "@nestjs/jwt";
-import { JwtPayload } from "../../types/jwt";
+import { JwtPayload } from "../../../types/jwt";
 import { GoogleUserDto } from "./dto/google-user.dto";
 
 @Injectable()
