+++ server/src/domain/routines/routines.controller.ts
