+++ server/src/domain/sbd_statistics/entities/sbd_statistics.entity.ts
