+++ server/src/domain/users/entities/user.entity.ts
