+++ server/src/types/request.d.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { GoogleUserDto } from "../oauth/google-oauth/dto/google-user.dto";
+import { GoogleUserDto } from "../domain/oauth/google-oauth/dto/google-user.dto";
 
 export interface RequestWithUser extends Request {
   user: GoogleUserDto;
